refactor: OCR 호출 부분 Infra로 분리
