+++ src/main/java/org/sopt/carena/infrastructure/ocr/adapter/OcrAdapter.java
@@ -1,8 +1,9 @@
-package org.sopt.carena.healthreport.adapter.out.web.ocr;
+package org.sopt.carena.infrastructure.ocr.adapter;
 
-import org.sopt.carena.healthreport.domain.value.ocr.OcrResponse;
+import org.sopt.carena.infrastructure.ocr.adapter.client.OcrClient;
+import org.sopt.carena.infrastructure.ocr.dto.OcrResponse;
 import org.sopt.carena.healthreport.application.dto.command.ExtractTextCommand;
-import org.sopt.carena.healthreport.application.port.out.OcrPort;
+import org.sopt.carena.infrastructure.ocr.port.OcrPort;
 import org.springframework.stereotype.Component;
 
 import lombok.RequiredArgsConstructor;
