+++ src/main/java/org/sopt/carena/infrastructure/ocr/port/OcrPort.java
@@ -1,6 +1,6 @@
-package org.sopt.carena.healthreport.application.port.out;
+package org.sopt.carena.infrastructure.ocr.port;
 
-import org.sopt.carena.healthreport.domain.value.ocr.OcrResponse;
+import org.sopt.carena.infrastructure.ocr.dto.OcrResponse;
 import org.sopt.carena.healthreport.application.dto.command.ExtractTextCommand;
 
 public interface OcrPort {
